requirements.txt
mcp==1.20.0
fastmcp==2.13.0.2
boto3==1.35.91
botocore==1.35.91
langchain-aws==0.2.9
python-dotenv==1.0.0
httpx==0.28.1
pydantic==2.12.3
pytest==8.4.2
black==24.10.0
pre-commit==4.3.0
arxiv==2.2.0
PyMuPDF==1.26.5